@@ -57,7 +57,7 @@ public class MonteMediaPlayer extends AbstractMediaPlayer {
     @Override
     public Duration getFrameAfter(Duration timestamp) {
         Rational t = engine.getFrameAfter(Rational.valueOf((long) timestamp.toMillis(), 1000));
-        Duration seconds = Duration.millis(t.multiply(1000).doubleValue());
+        Duration seconds = Duration.seconds(t.doubleValue());
         if (Math.abs(timestamp.toMillis() - seconds.toMillis()) < 0.01) {
             t = engine.getFrameAfter(t);
             seconds = Duration.millis(t.multiply(1000).doubleValue());
@@ -68,7 +68,7 @@ public class MonteMediaPlayer extends AbstractMediaPlayer {
     @Override
     public Duration getFrameBefore(Duration timestamp) {
         Rational t = engine.getFrameBefore(Rational.valueOf((long) timestamp.toMillis(), 1000));
-        Duration seconds = Duration.millis(t.multiply(1000).doubleValue());
+        Duration seconds = Duration.seconds(t.doubleValue());
         /*
         if(Math.abs(timestamp.toMillis()-seconds.toMillis())<0.01){
             t=engine.getFrameBefore(t);
