@@ -7,13 +7,14 @@ package org.monte.media.qtff;
 import org.monte.media.av.Format;
 import org.monte.media.av.FormatKeys.MediaType;
 import org.monte.media.math.Rational;
+import org.monte.media.qtff.atom.CompositeAtom;
+import org.monte.media.qtff.atom.WideDataAtom;
 
 import javax.imageio.stream.ImageOutputStream;
 import java.awt.image.IndexColorModel;
 import java.io.IOException;
 import java.time.Instant;
 import java.util.ArrayList;
-import java.util.LinkedList;
 
 /**
  * This is the base class for low-level QuickTime stream IO.
@@ -514,243 +515,6 @@ public class AbstractQTFFMovieStream {
             this.trackDuration = trackDuration;
             this.mediaTime = mediaTime;
             this.mediaRate = mediaRate;
-        }
-    }
-
-    /**
-     * Atom base class.
-     */
-    protected abstract class Atom {
-
-        /**
-         * The type of the atom. A String with the length of 4 characters.
-         */
-        protected String type;
-        /**
-         * The offset of the atom relative to the start of the
-         * ImageOutputStream.
-         */
-        protected long offset;
-
-        /**
-         * Creates a new Atom at the current position of the ImageOutputStream.
-         *
-         * @param type The type of the atom. A string with a length of 4
-         *             characters.
-         */
-        public Atom(String type, long offset) {
-            this.type = type;
-            this.offset = offset;
-        }
-
-        /**
-         * Writes the atom to the ImageOutputStream and disposes it.
-         */
-        public abstract void finish() throws IOException;
-
-        /**
-         * Returns the size of the atom including the size of the atom header.
-         *
-         * @return The size of the atom.
-         */
-        public abstract long size();
-    }
-
-    /**
-     * A CompositeAtom contains an ordered list of Atoms.
-     */
-    protected class CompositeAtom extends DataAtom {
-
-        protected LinkedList<Atom> children;
-
-        /**
-         * Creates a new CompositeAtom at the current position of the
-         * ImageOutputStream.
-         *
-         * @param type The type of the atom.
-         */
-        public CompositeAtom(String type) throws IOException {
-            super(type);
-            children = new LinkedList<>();
-        }
-
-        public void add(Atom child) throws IOException {
-            if (children.size() > 0) {
-                children.get(children.size() - 1).finish();
-            }
-            children.add(child);
-        }
-
-        /**
-         * Writes the atom and all its children to the ImageOutputStream and
-         * disposes of all resources held by the atom.
-         *
-         * @throws java.io.IOException
-         */
-        @Override
-        public void finish() throws IOException {
-            if (!finished) {
-                if (size() > 0xffffffffL) {
-                    throw new IOException("CompositeAtom \"" + type + "\" is too large: " + size());
-                }
-
-                long pointer = getRelativeStreamPosition();
-                seekRelative(offset);
-
-                QTFFImageOutputStream headerData = new QTFFImageOutputStream(out);
-                headerData.writeInt((int) size());
-                headerData.writeType(type);
-                for (Atom child : children) {
-                    child.finish();
-                }
-                seekRelative(pointer);
-                finished = true;
-            }
-        }
-
-        @Override
-        public long size() {
-            long length = 8 + data.length();
-            for (Atom child : children) {
-                length += child.size();
-            }
-            return length;
-        }
-    }
-
-    /**
-     * Data Atom.
-     */
-    protected class DataAtom extends Atom {
-
-        protected QTFFImageOutputStream data;
-        protected boolean finished;
-
-        /**
-         * Creates a new DataAtom at the current position of the
-         * ImageOutputStream.
-         *
-         * @param type The type name of the atom.
-         */
-        public DataAtom(String type) throws IOException {
-            super(type, getRelativeStreamPosition());
-            out.writeLong(0); // make room for the atom header
-            data = new QTFFImageOutputStream(out);
-        }
-
-        @Override
-        public void finish() throws IOException {
-            if (!finished) {
-                long sizeBefore = size();
-
-                if (size() > 0xffffffffL) {
-                    throw new IOException("DataAtom \"" + type + "\" is too large: " + size());
-                }
-
-                long pointer = getRelativeStreamPosition();
-                seekRelative(offset);
-
-                QTFFImageOutputStream headerData = new QTFFImageOutputStream(out);
-                headerData.writeInt((int) size());
-                headerData.writeType(type);
-                seekRelative(pointer);
-                finished = true;
-                long sizeAfter = size();
-                if (sizeBefore != sizeAfter) {
-                    System.err.println("size mismatch " + sizeBefore + ".." + sizeAfter);
-                }
-            }
-        }
-
-        /**
-         * Returns the offset of this atom to the beginning of the random access
-         * file
-         */
-        public long getOffset() {
-            return offset;
-        }
-
-        public QTFFImageOutputStream getOutputStream() {
-            if (finished) {
-                throw new IllegalStateException("DataAtom is finished");
-            }
-            return data;
-        }
-
-        @Override
-        public long size() {
-            return 8 + data.length();
-        }
-    }
-
-    /**
-     * WideDataAtom can grow larger then 4 gigabytes.
-     */
-    protected class WideDataAtom extends Atom {
-
-        protected QTFFImageOutputStream data;
-        protected boolean finished;
-
-        /**
-         * Creates a new DataAtom at the current position of the
-         * ImageOutputStream.
-         *
-         * @param type The type of the atom.
-         */
-        public WideDataAtom(String type) throws IOException {
-            super(type, getRelativeStreamPosition());
-            out.writeLong(0); // make room for the atom header
-            out.writeLong(0); // make room for the atom header
-            data = new QTFFImageOutputStream(out) {
-                @Override
-                public void flush() throws IOException {
-                    // DO NOT FLUSH UNDERLYING STREAM!
-                }
-            };
-        }
-
-        @Override
-        public void finish() throws IOException {
-            if (!finished) {
-                long pointer = getRelativeStreamPosition();
-                seekRelative(offset);
-
-                QTFFImageOutputStream headerData = new QTFFImageOutputStream(out);
-                long finishedSize = size();
-                if (finishedSize <= 0xffffffffL) {
-                    headerData.writeInt(8);
-                    headerData.writeType("wide");
-                    headerData.writeInt((int) (finishedSize - 8));
-                    headerData.writeType(type);
-                } else {
-                    headerData.writeInt(1); // special value for extended size atoms
-                    headerData.writeType(type);
-                    headerData.writeLong(finishedSize - 8);
-                }
-
-                seekRelative(pointer);
-                finished = true;
-            }
-        }
-
-        /**
-         * Returns the offset of this atom to the beginning of the random access
-         * file
-         */
-        public long getOffset() {
-            return offset;
-        }
-
-        public QTFFImageOutputStream getOutputStream() {
-            if (finished) {
-                throw new IllegalStateException("Atom is finished");
-            }
-            return data;
-        }
-
-        @Override
-        public long size() {
-            return 16 + data.length();
         }
     }
 
