@@ -10,10 +10,15 @@ import org.monte.media.av.FormatKeys.MediaType;
 import org.monte.media.av.codec.video.AbstractVideoCodec;
 import org.monte.media.io.ByteArrayImageOutputStream;
 import org.monte.media.util.ArrayUtil;
+import org.monte.media.util.ByteArrays;
 
 import javax.imageio.stream.ImageOutputStream;
 import java.awt.Rectangle;
 import java.awt.image.BufferedImage;
+import java.awt.image.ColorModel;
+import java.awt.image.DataBufferByte;
+import java.awt.image.DataBufferInt;
+import java.awt.image.DataBufferUShort;
 import java.awt.image.WritableRaster;
 import java.io.IOException;
 import java.io.OutputStream;
@@ -29,7 +34,9 @@ import static org.monte.media.av.codec.video.VideoFormatKeys.DataClassKey;
 import static org.monte.media.av.codec.video.VideoFormatKeys.DepthKey;
 import static org.monte.media.av.codec.video.VideoFormatKeys.ENCODING_BUFFERED_IMAGE;
 import static org.monte.media.av.codec.video.VideoFormatKeys.ENCODING_QUICKTIME_RAW;
+import static org.monte.media.av.codec.video.VideoFormatKeys.FixedFrameRateKey;
 import static org.monte.media.av.codec.video.VideoFormatKeys.HeightKey;
+import static org.monte.media.av.codec.video.VideoFormatKeys.PaletteKey;
 import static org.monte.media.av.codec.video.VideoFormatKeys.WidthKey;
 
 /**
@@ -60,8 +67,19 @@ public class RawCodec extends AbstractVideoCodec {
         super(new Format[]{
                         new Format(MediaTypeKey, MediaType.VIDEO, MimeTypeKey, MIME_JAVA,
                                 EncodingKey, ENCODING_BUFFERED_IMAGE), //
+                        new Format(MediaTypeKey, MediaType.VIDEO, MimeTypeKey, MIME_QUICKTIME,
+                                EncodingKey, ENCODING_QUICKTIME_RAW, DataClassKey, byte[].class,
+                                FixedFrameRateKey, true, DepthKey, 16), //
+                        new Format(MediaTypeKey, MediaType.VIDEO, MimeTypeKey, MIME_QUICKTIME,
+                                EncodingKey, ENCODING_QUICKTIME_RAW, DataClassKey, byte[].class,
+                                FixedFrameRateKey, true, DepthKey, 8), //
+                        new Format(MediaTypeKey, MediaType.VIDEO, MimeTypeKey, MIME_QUICKTIME,
+                                EncodingKey, ENCODING_QUICKTIME_RAW, DataClassKey, byte[].class,
+                                FixedFrameRateKey, true, DepthKey, 24), //
                 },
                 new Format[]{
+                        new Format(MediaTypeKey, MediaType.VIDEO, MimeTypeKey, MIME_JAVA,
+                                EncodingKey, ENCODING_BUFFERED_IMAGE, FixedFrameRateKey, true), //
                         new Format(MediaTypeKey, MediaType.VIDEO, MimeTypeKey, MIME_QUICKTIME,
                                 EncodingKey, ENCODING_QUICKTIME_RAW, DataClassKey, byte[].class, DepthKey, 8), //
                         new Format(MediaTypeKey, MediaType.VIDEO, MimeTypeKey, MIME_QUICKTIME,
@@ -107,8 +125,75 @@ public class RawCodec extends AbstractVideoCodec {
         }
     }
 
+    public void readKey8(byte[] in, int offset, int length, BufferedImage img) {
+        DataBufferByte buf = (DataBufferByte) img.getRaster().getDataBuffer();
+        WritableRaster raster = img.getRaster();
+        int scanlineStride = raster.getSampleModel().getWidth();
+        Rectangle r = raster.getBounds();
+        r.x -= raster.getSampleModelTranslateX();
+        r.y -= raster.getSampleModelTranslateY();
+
+        int h = img.getHeight();
+        int w = img.getWidth();
+        int i = offset;
+        int xy = 0;
+        byte[] out = buf.getData();
+        for (int y = 0; y < h; y++) {
+            System.arraycopy(in, i, out, xy, w);
+            i += w;
+            xy += scanlineStride;
+        }
+    }
+
+    public void readKey16(byte[] in, int offset, int length, BufferedImage img) {
+        DataBufferUShort buf = (DataBufferUShort) img.getRaster().getDataBuffer();
+        WritableRaster raster = img.getRaster();
+        int scanlineStride = raster.getSampleModel().getWidth();
+        Rectangle r = raster.getBounds();
+        r.x -= raster.getSampleModelTranslateX();
+        r.y -= raster.getSampleModelTranslateY();
+
+        int h = img.getHeight();
+        int w = img.getWidth();
+        int i = offset;
+        int xy = 0;
+        short[] out = buf.getData();
+        for (int y = 0; y < h; y++) {
+            for (int k = 0, k2 = 0; k < w; k++, k2 += 2) {
+                out[xy + k] = ByteArrays.getShortBE(in, i + k2);
+            }
+            i += w * 2;
+            xy += scanlineStride;
+        }
+    }
+
+    public void readKey24(byte[] in, int offset, int length, BufferedImage img) {
+        DataBufferInt buf = (DataBufferInt) img.getRaster().getDataBuffer();
+        WritableRaster raster = img.getRaster();
+        int scanlineStride = raster.getSampleModel().getWidth();
+        Rectangle r = raster.getBounds();
+        r.x -= raster.getSampleModelTranslateX();
+        r.y -= raster.getSampleModelTranslateY();
+
+        int h = img.getHeight();
+        int w = img.getWidth();
+        int i = offset;
+        int xy = 0;
+        int[] out = buf.getData();
+        for (int y = 0; y < h; y++) {
+            for (int k = 0, k3 = 0; k < w; k++, k3 += 3) {
+                out[xy + k] = 0xff000000//Alpha
+                        | ((in[i + k3] & 0xff) << 16)//Red
+                        | ((in[i + k3 + 1] & 0xff) << 8)//Green
+                        | ((in[i + k3 + 2] & 0xff));//Blue
+            }
+            i += w * 3;
+            xy += scanlineStride;
+        }
+    }
+
     /**
-     * Encodes a 24-bit key frame.
+     * Encodes a 16-bit key frame.
      *
      * @param out            The output stream.
      * @param data           The image data.
@@ -121,12 +206,11 @@ public class RawCodec extends AbstractVideoCodec {
             throws IOException {
 
         // Write the samples
-        byte[] bytes = new byte[width * 2]; // holds a scanline of raw image data onlyWith 3 channels of 32 bit data
+        byte[] bytes = new byte[width * 2]; // holds a scanline of raw image data onlyWith 3 channels of 16 bit data
         for (int xy = offset, ymax = offset + height * scanlineStride; xy < ymax; xy += scanlineStride) {
             for (int x = 0, i = 0; x < width; x++, i += 2) {
-                int pixel = data[xy + x];
-                bytes[i] = (byte) (pixel >> 8);
-                bytes[i + 1] = (byte) (pixel);
+                short pixel = data[xy + x];
+                ByteArrays.setShortBE(bytes, i, pixel);
             }
             out.write(bytes, 0, bytes.length);
         }
@@ -159,7 +243,7 @@ public class RawCodec extends AbstractVideoCodec {
     }
 
     /**
-     * Encodes a 24-bit key frame.
+     * Encodes a 32-bit key frame.
      *
      * @param out            The output stream.
      * @param data           The image data.
@@ -176,10 +260,7 @@ public class RawCodec extends AbstractVideoCodec {
         for (int xy = offset, ymax = offset + height * scanlineStride; xy < ymax; xy += scanlineStride) {
             for (int x = 0, i = 0; x < width; x++, i += 4) {
                 int pixel = data[xy + x];
-                bytes[i] = (byte) (pixel >> 24);
-                bytes[i + 1] = (byte) (pixel >> 16);
-                bytes[i + 2] = (byte) (pixel >> 8);
-                bytes[i + 3] = (byte) (pixel);
+                ByteArrays.setIntBE(bytes, i, pixel);
             }
             out.write(bytes, 0, bytes.length);
         }
@@ -211,6 +292,78 @@ public class RawCodec extends AbstractVideoCodec {
 
     @Override
     public int process(Buffer in, Buffer out) {
+        if (outputFormat.get(EncodingKey) == ENCODING_BUFFERED_IMAGE) {
+            return decode(in, out);
+        } else {
+            return encode(in, out);
+        }
+    }
+
+    public int decode(Buffer in, Buffer out) {
+        out.setMetaTo(in);
+        out.format = outputFormat;
+        if (in.isFlag(DISCARD)) {
+            return CODEC_OK;
+        }
+
+        out.sampleCount = 1;
+        BufferedImage img = null;
+
+        int imgType;
+        ColorModel cm;
+        switch (inputFormat.get(DepthKey)) {
+            case 8 -> {
+                cm = inputFormat.get(PaletteKey);
+                imgType = BufferedImage.TYPE_BYTE_INDEXED;
+            }
+            case 16 -> {
+                cm = inputFormat.get(PaletteKey);
+                imgType = BufferedImage.TYPE_USHORT_555_RGB;
+            }
+            default -> {
+                cm = null;
+                imgType = BufferedImage.TYPE_INT_RGB;
+            }
+        }
+        ;
+
+        int width = inputFormat.get(WidthKey);
+        int height = inputFormat.get(HeightKey);
+        if (out.data instanceof BufferedImage) {
+            img = (BufferedImage) out.data;
+            if (img != null && img.getWidth() != width
+                    || img.getHeight() != height
+                    || img.getType() != imgType) {
+                img = null;
+            }
+        }
+        if (img == null) {
+            if (cm != null) {
+                img = new BufferedImage(
+                        cm,
+                        cm.createCompatibleWritableRaster(width, height), false, null);
+            } else {
+                img = new BufferedImage(width, height, imgType);
+            }
+        }
+        out.data = img;
+
+        switch (inputFormat.get(DepthKey)) {
+            case 8:
+                readKey8((byte[]) in.data, in.offset, in.length, img);
+                break;
+            case 16:
+                readKey16((byte[]) in.data, in.offset, in.length, img);
+                break;
+            case 24:
+            default:
+                readKey24((byte[]) in.data, in.offset, in.length, img);
+                break;
+        }
+        return CODEC_OK;
+    }
+
+    public int encode(Buffer in, Buffer out) {
         out.setMetaTo(in);
         if (in.isFlag(DISCARD)) {
             return CODEC_OK;
